Remove naive solver (#3161)
# Description
The naive solver is extremely basic and never wins any auctions anymore.
Also AFAICT it's no longer used in e2e tests either so by now it's just
dead weight.
Sparked by this
[thread](https://cowservices.slack.com/archives/C036JAGRQ04/p1733911881085109)

# Changes
Removes:
* original algorithm in legacy code
* wrapper logic in `solvers` crate
* config code
* test cases
* documentation

Also in many cases we had an enum with 2 variants (`baseline` and
`naive`). Since we now only have 1 variant left I collapsed all the
logic and got rid of a few modules which are no longer necessary.
