--- crates/solver/src/solver/mod.rs
@@ -1,25 +1,0 @@
-use {
-    anyhow::anyhow,
-    std::{fmt::Debug, str::FromStr},
-};
-
-mod baseline_solver;
-pub mod naive_solver;
-
-// Wrapper type for AWS ARN identifiers
-#[derive(Debug, Clone)]
-pub struct Arn(pub String);
-
-impl FromStr for Arn {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        // Could be more strict here, but this should suffice to catch unintended
-        // configuration mistakes
-        if s.starts_with("arn:aws:kms:") {
-            Ok(Self(s.to_string()))
-        } else {
-            Err(anyhow!("Invalid ARN identifier: {}", s))
-        }
-    }
-}
