+++ crates/solvers/src/boundary/mod.rs
@@ -3,6 +3,5 @@
 
 pub mod baseline;
 pub mod liquidity;
-pub mod naive;
 
 pub type Result<T> = anyhow::Result<T>;
