+++ crates/solvers/src/tests/cases/buy_order_rounding.rs
