+++ crates/solvers/src/tests/cases/direct_swap.rs
