+++ crates/solvers/src/tests/cases/internalization.rs
