+++ crates/solvers/src/tests/cases/mod.rs
