+++ crates/solvers/src/tests/cases/partial_fill.rs
