--- crates/solvers/src/tests/naive/limit_order_price.rs
@@ -1,70 +1,0 @@
-//! This test verifies that the limit order's limit price is respected after
-//! surplus fees are taken from the order.
-
-use {crate::tests, serde_json::json};
-
-#[tokio::test]
-async fn test() {
-    let engine = tests::SolverEngine::new("naive", tests::Config::None).await;
-
-    let solution = engine
-        .solve(json!({
-            "id": "1",
-            "tokens": {},
-            "orders": [
-                {
-                    "uid": "0x2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a\
-                              2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a\
-                              2a2a2a2a",
-                    "sellToken": "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-                    "buyToken": "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
-                    "sellAmount": "22397494",
-                    "fullSellAmount": "1000000000000000000",
-                    "buyAmount": "18477932550000000",
-                    "fullBuyAmount": "1000000000000000000000",
-                    "feePolicies": [],
-                    "validTo": 0,
-                    "kind": "sell",
-                    "owner": "0x5b1e2c2762667331bc91648052f646d1b0d35984",
-                    "partiallyFillable": false,
-                    "preInteractions": [],
-                    "postInteractions": [],
-                    "sellTokenSource": "erc20",
-                    "buyTokenDestination": "erc20",
-                    "class": "limit",
-                    "appData": "0x6000000000000000000000000000000000000000000000000000000000000007",
-                    "signingScheme": "presign",
-                    "signature": "0x",
-                },
-            ],
-            "liquidity": [
-                {
-                    "kind": "constantProduct",
-                    "tokens": {
-                        "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48": {
-                            "balance": "36338096110368"
-                        },
-                        "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2": {
-                            "balance": "30072348537379906026018"
-                        }
-                    },
-                    "fee": "0.003",
-                    "id": "0",
-                    "address": "0x0000000000000000000000000000000000000000",
-                    "router": "0xffffffffffffffffffffffffffffffffffffffff",
-                    "gasEstimate": "110000"
-                },
-            ],
-            "effectiveGasPrice": "15000000000",
-            "deadline": "2106-01-01T00:00:00.000Z",
-            "surplusCapturingJitOrderOwners": []
-        }))
-        .await;
-
-    assert_eq!(
-        solution,
-        json!({
-            "solutions": []
-        }),
-    );
-}
