--- crates/solvers/src/tests/naive/mod.rs
@@ -1,8 +1,0 @@
-mod extract_deepest_pool;
-mod filters_out_of_price_orders;
-mod limit_order_price;
-mod matches_orders;
-mod reserves_too_small;
-mod rounds_prices_in_favour_of_traders;
-mod swap_less_than_reserves;
-mod without_pool;
